+++ contracts/include/ups-web4.hpp
@@ -28,7 +28,7 @@
 
 
     // --- Return a Series --- //
-    vector<uint64_t> getSeries(const string& seriesName){
+    vector<uint32_t> getSeries(const string& seriesName){
         if (seriesName == "FIBONACCI") {
             return series::FIBONACCI;
         } else if (seriesName == "SILVER") {
